Enhanced maintainability and readability of the codebase by adhering to modular design principles.
