+++ PanelGeneMapper/modules/logging_utils.py
@@ -1,0 +1,14 @@
+import logging
+
+def setup_logging():
+    """
+    Configure logging settings.
+    """
+    logging.basicConfig(
+        level=logging.INFO,
+        format="%(asctime)s - %(levelname)s - %(message)s",
+        handlers=[
+            logging.FileHandler("retrieve_data.log"),
+            logging.StreamHandler()
+        ]
+    )
